pyproject.toml
[tool.poetry]
name = "chat-with-your-pdf-multimodal-rag-system"
version = "0.1.0"
description = ""
authors = ["Omar Elsherif <omarelsherif010@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
tiktoken = "^0.8.0"
pillow = "^11.1.0"
lxml = "^5.3.0"
chromadb = "^0.6.3"
langchain = "^0.3.17"
langchain-community = "^0.3.16"
langchain-openai = "^0.3.3"
langchain-groq = "^0.2.4"
python-dotenv = "^1.0.1"
python-magic-bin = "^0.4.14"
langchain-unstructured = "^0.1.6"
unstructured = {extras = ["pdf"], version = "^0.16.17"}
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
pypdf = "^5.2.0"
pymupdf = "^1.25.2"
pandas = "^2.2.3"
streamlit = "^1.42.0"
pydantic-settings = "^2.7.1"
langchain-chroma = "^0.2.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
